tests restructure and workflow
